Fixes #179134. Adds flag suppressSuggestions to inline completion result.
