+++ src/vs/editor/common/languages.ts
@@ -691,6 +691,8 @@ export interface InlineCompletions<TItem extends InlineCompletion = InlineComple
 	 * A list of commands associated with the inline completions of this list.
 	 */
 	readonly commands?: Command[];
+
+	readonly suppressSuggestions?: boolean | undefined;
 }
 
 export interface InlineCompletionsProvider<T extends InlineCompletions = InlineCompletions> {
